→ Moves: Error mapper functions to State itself
- Removes: the mapper from the viewmodel
