+++ WordPress/src/main/java/org/wordpress/android/ui/blaze/blazecampaigns/campaignlisting/CampaignListingUIModelMapper.kt
@@ -37,31 +37,4 @@ class CampaignListingUIModelMapper @Inject constructor(
     private fun convertToDollars(budgetCents: Long): UiString {
         return UiString.UiStringText("$" + (budgetCents / CENTS_IN_DOLLARS).toString())
     }
-
-    fun toNoCampaignsError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
-        title = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_message_title),
-        description = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_message_description),
-        button = CampaignListingUiState.Error.ErrorButton(
-            text = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_button_text),
-            click = buttonClick
-        )
-    )
-
-    fun toNoNetworkError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
-        title = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_title),
-        description = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_description),
-        button = CampaignListingUiState.Error.ErrorButton(
-            text = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_button_text),
-            click = buttonClick
-        )
-    )
-
-    fun toGenericError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
-        title = UiString.UiStringRes(R.string.campaign_listing_page_error_title),
-        description = UiString.UiStringRes(R.string.campaign_listing_page_error_description),
-        button = CampaignListingUiState.Error.ErrorButton(
-            text = UiString.UiStringRes(R.string.campaign_listing_page_error_button_text),
-            click = buttonClick
-        )
-    )
 }
