+++ WordPress/src/main/java/org/wordpress/android/ui/blaze/blazecampaigns/campaignlisting/CampaignListingUiState.kt
@@ -1,5 +1,6 @@
 package org.wordpress.android.ui.blaze.blazecampaigns.campaignlisting
 
+import org.wordpress.android.R
 import org.wordpress.android.ui.mysite.cards.blaze.CampaignStatus
 import org.wordpress.android.ui.utils.UiString
 
@@ -26,6 +27,35 @@ sealed class CampaignListingUiState {
         data class PagingDetails(
             val loadMoreFunction: () -> Unit = { },
             val loadingNext: Boolean = false
+        )
+    }
+
+    companion object {
+        fun toNoCampaignsError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
+            title = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_message_title),
+            description = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_message_description),
+            button = Error.ErrorButton(
+                text = UiString.UiStringRes(R.string.campaign_listing_page_no_campaigns_button_text),
+                click = buttonClick
+            )
+        )
+
+        fun toNoNetworkError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
+            title = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_title),
+            description = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_description),
+            button = Error.ErrorButton(
+                text = UiString.UiStringRes(R.string.campaign_listing_page_no_network_error_button_text),
+                click = buttonClick
+            )
+        )
+
+        fun toGenericError(buttonClick: () -> Unit) = CampaignListingUiState.Error(
+            title = UiString.UiStringRes(R.string.campaign_listing_page_error_title),
+            description = UiString.UiStringRes(R.string.campaign_listing_page_error_description),
+            button = Error.ErrorButton(
+                text = UiString.UiStringRes(R.string.campaign_listing_page_error_button_text),
+                click = buttonClick
+            )
         )
     }
 }
